moved request to separate file
