+++ src/scripts/main.js
@@ -1,25 +1,8 @@
 'use strict';
 
-const BASE_URL = 'https://mate-academy.github.io/phone-catalogue-static/api';
-const ENDPOINTS = {
-  phones: '/phones.json',
-  phoneById: (id) => `/phones/${id}.json`,
-};
+const { ENDPOINTS, request } = require('./requests');
 const body = document.querySelector('body');
 const list = document.createElement('ul');
-
-const request = (url) => {
-  return fetch(`${BASE_URL}${url}`)
-    .then(response => {
-      if (!response.ok) {
-        setTimeout(() => Promise.reject(
-          new Error(`${response.status} - ${response.statusText}`)
-        ), 5000);
-      }
-
-      return response.json();
-    });
-};
 
 const getPhoneDetails = (idArr) => {
   return Promise.all(idArr.map(id => request(ENDPOINTS.phoneById(id))));
